Add roundtrip conversion tests
